added a search input box but its not hooked up to anything
