+++ src/App.js
@@ -3,7 +3,7 @@ import { Route, Switch } from "react-router-dom";
 import "./App.css";
 // * Component Imports
 import RoutesTree from "./Components/RoutesTree";
-import Header from "./Components/SharedComponents/Header";
+import Header from "./Components/HeaderComponents/Header";
 import Footer from "./Components/SharedComponents/Footer";
 
 function App() {
