+++ src/Components/NavBarComponents/SearchBar.jsx
@@ -1,0 +1,15 @@
+import React from "react";
+
+const SearchBar = () => {
+  return (
+    <div>
+      <input
+        class="input is-rounded"
+        type="text"
+        placeholder="Just an input box"
+      />
+    </div>
+  );
+};
+
+export default SearchBar;
