--- src/Components/SharedComponents/Header.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-// * Component Imports
-import NavBar from "./NavBar";
-
-const Header = () => {
-  return (
-    <div className="header">
-      <NavBar />
-    </div>
-  );
-};
-
-export default Header;
