--- src/Components/SharedComponents/SearchBar.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const SearchBar = () => {
-  return <div>I am a search bar</div>;
-};
-
-export default SearchBar;
